backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
aiogram 